feat: ship py.typed; type-check src with mypy in CI (#18)
* feat: ship py.typed; type-check src with mypy in CI

The package now declares itself typed (PEP 561) and CI's lint job runs
mypy over src/ (check_untyped_defs, warn_unreachable, warn_unused_ignores;
barcode/pystrich have no stubs and are ignored, qrcode uses types-qrcode).

Fixes surfaced by the checker, none of which change rendered output
(golden suite unchanged):
- plot: `data: []` reached math.ceil(None) and raised TypeError; it is now
  a RenderError naming the problem. The same-column collapse loop no
  longer carries an Optional last_x.
- registry: real forward reference to RenderState under TYPE_CHECKING.
- context: layout_engine is typed ImageFont.Layout | None.
- state: pos_y is float (Pillow's textbbox edges are floats).
- dither: the optional numpy module is typed ModuleType | None so both
  branches type-check; the pure ordered path reads/writes bytes like the
  error-diffusion path (identical output, asserted by the pure-vs-numpy
  and pure-vs-golden tests).
- codes/media: Image.Resampling enum; qrcode builds a PilImage explicitly
  and takes .get_image(); barcode sizing branches are exhaustive.
- The wheel check in CI asserts py.typed is packaged.


* chore: ignore .mypy_cache


---------

+++ src/imagespec/state.py
@@ -29,4 +29,4 @@ class RenderState:
     canvas_width: int
     canvas_height: int
     context: RenderContext
-    pos_y: int = 0
+    pos_y: float = 0
